🎨 Modularize installDependencies and requireHook
