+++ lib/injection.js
@@ -1,27 +1,12 @@
 const fs = require('fs');
-const crypto = require('crypto');
-const { execSync } = require('child_process');
 const JSON5 = require('json5');
 const path = require('path');
+
+const requireHook = require('./utils/requireHook');
+const installDependencies = require('./utils/installDependencies');
+
 const tsconfigDir = process.cwd();
 const tsconfigPath = path.resolve(process.cwd(), 'tsconfig.json');
-
-function requireHook(hook) {
-	if (hook.endsWith('.js')) {
-		try {
-			new URL(hook);
-			const hashPath = path.resolve(__dirname, '../hooks', `${crypto.createHash('md5').update(hook).digest('hex')}.js`);
-			if (!fs.existsSync(hashPath)) {
-				execSync(`curl ${hook} -o ${hashPath} -s`);
-			}
-			return require(hashPath);
-		} catch (error) {
-			return require(path.resolve(tsconfigDir, hook));
-		}
-	} else {
-		return require(path.resolve(__dirname, '../hooks', hook));
-	}
-}
 
 if (fs.existsSync(tsconfigPath)) {
 	const tsconfigStr = fs.readFileSync(tsconfigPath);
@@ -50,9 +35,8 @@ if (fs.existsSync(tsconfigPath)) {
 		hookModules.push(hookModule);
 	}
 
-	if (dependencies.length) {
-		execSync(`npm i ${Array.from(new Set(dependencies)).join(' ')}`, { cwd: path.resolve(__dirname, '..') });
-	}
+	// Install dependencies
+	installDependencies(dependencies);
 
 	if (config.length) {
 		const copyTSConfig = JSON5.parse(tsconfigStr);
@@ -68,7 +52,7 @@ if (fs.existsSync(tsconfigPath)) {
 			deps[dep] = require(dep);
 			return deps;
 		}, {}),
-		config: tsconfig,
+		tsconfig,
 		tsconfigDir,
 		tsconfigPath,
 		args: arguments
