+++ lib/utils/installDependencies.js
@@ -1,0 +1,5 @@
+module.exports = function(dependencies) {
+  if (dependencies.length) {
+		execSync(`npm i ${Array.from(new Set(dependencies)).join(' ')}`, { cwd: path.resolve(__dirname, '..') });
+	}
+}
